requirements.txt
streamlit
huggingface_hub
pyannote.audio
pyannote.core
whisper